Added handler to prorate/not prorate swaps between plans
