+++ src/BillingPortal.php
@@ -3,18 +3,12 @@
 namespace RenokiCo\BillingPortal;
 
 use Closure;
-use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\Request;
-use RenokiCo\CashierRegister\Plan;
 
 class BillingPortal
 {
-    /**
-     * The callback to sync quotas for an user.
-     *
-     * @var Closure|null
-     */
-    protected static $syncQuotasResolver;
+    use Traits\ResolvesQuotas;
+    use Traits\ResolvesStripeCheckout;
 
     /**
      * The closure that will be called to retrieve
@@ -25,47 +19,30 @@ class BillingPortal
     protected static $billable;
 
     /**
-     * The closure that will be called to get
-     * the right Stripe Checkout parameters to call.
+     * Wether the proration should occur when swapping between plans.
      *
-     * @var null|Closure
+     * @var bool
      */
-    protected static $stripeCheckoutOptions;
+    protected static $proratesOnSwap = true;
 
     /**
-     * The closure that will be called to modify
-     * the Stripe Checkout flow.
+     * Don't prorate on swapping.
      *
-     * @var null|Closure
-     */
-    protected static $stripeCheckoutResolver;
-
-    /**
-     * Register a method that will run when the
-     * subscription updates, in order to sync the quotas.
-     *
-     * @param  \Closure  $callback
      * @return void
      */
-    public static function resolveQuotasSync(Closure $callback)
+    public static function dontProrateOnSwap()
     {
-        static::$syncQuotasResolver = $callback;
+        static::$proratesOnSwap = false;
     }
 
     /**
-     * Run the syncing quotas callback.
+     * Wether the proration should occur when swapping.
      *
-     * @param  \Illuminate\Database\Eloquent\Model  $billable
-     * @param  \Illuminate\Database\Eloquent\Model  $subscription
-     * @return void
+     * @return bool
      */
-    public static function syncQuotas(Model $billable, Model $subscription)
+    public static function proratesOnSwap(): bool
     {
-        if (static::$syncQuotasResolver) {
-            $callback = static::$syncQuotasResolver;
-
-            $callback($billable, $subscription);
-        }
+        return static::$proratesOnSwap;
     }
 
     /**
@@ -93,66 +70,5 @@ class BillingPortal
         return $closure
             ? $closure($request)
             : $request->user();
-    }
-
-    /**
-     * Set the Stripe Checkout options computator.
-     *
-     * @param  \Closure  $callback
-     * @return void
-     */
-    public static function resolveStripeCheckoutOptions(Closure $callback)
-    {
-        static::$stripeCheckoutOptions = $callback;
-    }
-
-    /**
-     * Calculate the options for Stripe Checkout for
-     * a specific Billable mode, plan and subscription name.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  mixed  $billable
-     * @param  \RenokiCo\CashierRegister\Plan  $plan
-     * @param  string  $subscription
-     * @return array
-     */
-    public static function getStripeCheckoutOptions(Request $request, $billable, Plan $plan, string $subscription): array
-    {
-        $closure = static::$stripeCheckoutOptions;
-
-        return $closure
-            ? $closure($request, $billable, $plan, $subscription)
-            : [];
-    }
-
-    /**
-     * Set the Stripe Checkout interceptor.
-     *
-     * @param  \Closure  $callback
-     * @return void
-     */
-    public static function resolveStripeCheckout(Closure $callback)
-    {
-        static::$stripeCheckoutResolver = $callback;
-    }
-
-    /**
-     * Mutate the Stripe checkout for
-     * a specific Billable mode, plan and subscription name.
-     *
-     * @param  \Laravel\Cashier\SubscriptionBuilder  $checkout
-     * @param  \Illuminate\Http\Request  $request
-     * @param  mixed  $billable
-     * @param  \RenokiCo\CashierRegister\Plan  $plan
-     * @param  string  $subscription
-     * @return \Laravel\Cashier\SubscriptionBuilder
-     */
-    public static function mutateCheckout($checkout, Request $request, $billable, Plan $plan, string $subscription)
-    {
-        $closure = static::$stripeCheckoutResolver;
-
-        return $closure
-            ? $closure($checkout, $request, $billable, $plan, $subscription)
-            : $checkout;
     }
 }
